updated integrations routes
